AgricultureBlockchain/requirements.txt
setuptools
django==2.1.7
pymysql
pyaes
pbkdf2
python-dotenv 